feat: Long 타입에 memberCashReset 추가.

+++ src/main/java/io/cavia/trader/module/member/mypage/service/MyPageService.java
@@ -18,7 +18,7 @@ public interface MyPageService {
 
     public int changePassword(int id, String password, LocalDateTime passwordUpdatedAt);
 
-    public int resetCash(int id);
+    public int resetCash(Long id);
 
     public int deleteMember(Long id, String password);
 }
